// frontend/src/Enums/UserRoles.ts
enum UserRoles{
    Admin = 'admin',
    Finance = "finance",
    BranchManager = "branch_manager",
    Sells = "sells",
    Guest = "guest",
    System = "system",
    ProductionManager = "production_manager",
    AppBuilder = "app_builder",
    ChargeStationOperator = "charge_station_operator",
    ChargeStationOwner = "owner",
    EVDriver = "ev_driver"
}

export default UserRoles;